refactor(guardian): unify Cline/Junie/Cursor adapter main() bodies
EGC-539 audit (Finding 3): these three host adapters need a JSON
envelope on stdout (not a plain exit code), unlike the adapters that
already share runPlainExitCodeGuardianAdapter, so each one
independently reimplemented the same skeleton: read stdin -> check
truncated (identical message, copy-pasted three times) -> check ok ->
build guardian input -> run() -> check exitCode === 2 -> translate to
the host's own envelope shape.

Added runJsonEnvelopeGuardianAdapter to scripts/lib/adapter-stdin-json.js,
parameterized by buildGuardianInput (event -> Guardian input | null)
and respond(blocked, message) (host-specific envelope writer). Each
adapter now only supplies those two functions plus its own respond
shape; the truncation/ok/blocked control flow lives in one place.

Side effect: cursor-guardian-adapter.js was still using process.exit()
right after stdout.write(), the exact POSIX pipe-race pattern already
fixed for Cline/Junie in PR #1081 (a forced exit can race the flush
and truncate the response before the caller reads it). Consolidating
onto the shared helper's process.exitCode pattern fixes this adapter
too; added the same "full stdout present" regression test Cline/Junie
already had, which Cursor was missing.

Verified: full regression sweep across all 8 adapter test files plus
the shared module's own tests (90 tests total, 0 failures). No
behavior change for any adapter's actual allow/deny decisions or
output shape -- only the internal control flow moved.
