+++ scripts/hooks/junie-guardian-adapter.js
@@ -17,60 +17,38 @@
 'use strict';
 
 const { run } = require('./pre-bash-guardian-validate');
-const { readAdapterStdinJson } = require('../lib/adapter-stdin-json');
+const { runJsonEnvelopeGuardianAdapter } = require('../lib/adapter-stdin-json');
 
-function respond(decision, reason) {
+function buildGuardianInput(event) {
+  const command = event && typeof event === 'object' ? event.tool_input?.command : undefined;
+  if (!command) {
+    return null;
+  }
+  const guardianInput = { tool_name: 'Bash', tool_input: { command } };
+  if (event && typeof event.cwd === 'string') {
+    guardianInput.cwd = event.cwd;
+  }
+  return guardianInput;
+}
+
+function respond(blocked, reason) {
   // Sets exitCode and lets Node drain stdout naturally instead of forcing
   // process.exit() right after write(): on POSIX, stdout writes to a pipe
   // (which is what a hook subprocess always has) are asynchronous, so a
   // forced exit can race the write and truncate the response before Junie
   // reads it (cubic-dev-ai finding, PR #1081).
+  const decision = blocked ? 'deny' : 'allow';
   const payload = reason ? { decision, reason } : { decision };
-  process.exitCode = decision === 'deny' ? 2 : 0;
+  process.exitCode = blocked ? 2 : 0;
   process.stdout.write(JSON.stringify(payload));
 }
 
 function main() {
-  readAdapterStdinJson(({ ok, truncated, value }) => {
-    // Same fail-closed-on-truncation guard as the Cursor/Windsurf/Kiro
-    // adapters (cubic-dev-ai finding, PR #1073/#1074): a capped-length
-    // prefix can still happen to be syntactically valid JSON, so truncation
-    // is checked unconditionally, before `ok`.
-    if (truncated) {
-      respond('deny', 'EGC Guardian BLOCKED this command: the event payload exceeded the size ' +
-        'this validator can safely read, so it could not be parsed or validated. ' +
-        'Simplify the command.');
-      return;
-    }
-    if (!ok) {
-      respond('allow');
-      return;
-    }
-
-    const event = value;
-    const command = event && typeof event === 'object' ? event.tool_input?.command : undefined;
-    if (!command) {
-      respond('allow');
-      return;
-    }
-
-    const guardianInput = { tool_name: 'Bash', tool_input: { command } };
-    if (event && typeof event.cwd === 'string') {
-      guardianInput.cwd = event.cwd;
-    }
-
-    const result = run(guardianInput);
-    if (result.exitCode === 2) {
-      respond('deny', result.stderr || 'Blocked by the EGC Guardian.');
-      return;
-    }
-
-    respond('allow');
-  });
+  runJsonEnvelopeGuardianAdapter(buildGuardianInput, run, respond);
 }
 
 if (require.main === module) {
   main();
 }
 
-module.exports = {};
+module.exports = { buildGuardianInput };
